-Hide/show create buttons based on authState -Update edit and delete cell styles

+++ src/app/components/item/item.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { AuthService } from '../../core/auth.service';
 import { AngularFirestore, AngularFirestoreCollection, AngularFirestoreDocument } from "angularfire2/firestore";
 import { Observable } from 'rxjs'
 import { map } from 'rxjs/operators';
@@ -19,7 +20,8 @@ export class ItemComponent implements OnInit {
   selectedItem: Item;
   newItem: Item;
 
-  constructor(private afs: AngularFirestore) { 
+
+  constructor(private auth: AuthService, private afs: AngularFirestore) { 
     console.log('Constructing: ItemComponent with AngularFirestore...');
   }
 
@@ -49,22 +51,29 @@ export class ItemComponent implements OnInit {
   }
 
   addNewItem(): void {
-    this.newItem = {name: "", description: "", price: 0};
+    const currentUser = this.auth.getCurrentUser();
+    if(currentUser) {
+      this.newItem = {userId: currentUser.uid, name: "", description: "", price: 0};
+    }
     this.selectedItem = null;
   }
   
-
-
   updateItem(): void {
-    console.log("updating item...");
-    this.itemDoc = this.afs.doc(`items/${this.selectedItem.id}`);
-    this.itemDoc.update(this.selectedItem);
+    const currentUser = this.auth.getCurrentUser();
+    if(currentUser) {
+      console.log("updating item...");
+      this.itemDoc = this.afs.doc(`items/${this.selectedItem.id}`);
+      this.itemDoc.update(this.selectedItem);
+    }
   }
 
   deleteItem(item): void {
-    console.log("deleting item...");
-    this.itemDoc = this.afs.doc(`items/${item.id}`);
-    this.itemDoc.delete();
+    const currentUser = this.auth.getCurrentUser();
+    if(currentUser) {
+      console.log("deleting item...");
+      this.itemDoc = this.afs.doc(`items/${item.id}`);
+      this.itemDoc.delete();
+    }
   }
 
   addItem(): void {
